minor_project/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==5.0.6
django-crispy-forms==2.2
django-widget-tweaks==1.5.0
httplib2==0.22.0
idna==3.7
mysqlclient==2.2.4
oauth2==1.9.0.post1
oauthlib==3.2.2
pillow==10.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-openid==2.2.5
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
